refactor: reporting database models in individual files
